Redact country entity

+++ src/main/java/com/lashkevich/stores/entity/Country.java
@@ -6,7 +6,7 @@ public class Country {
     private long id;
     private String name;
 
-    public Country(int id, String name) {
+    public Country(long id, String name) {
         this.id = id;
         this.name = name;
     }
